requirements.txt
mock
pytest
flask
requests==2.22.0
psycopg2==2.8.6
python_dateutil==2.8.1
